debuggin new code

+++ src/Density/add_q_contribution_to_valence_density.cpp
@@ -14,9 +14,9 @@ void Density::add_q_contribution_to_valence_density(K_set& ks)
     pp_complex_density_matrix.zero();
     
     /* add k-point contribution */
-    for (int ikloc = 0; ikloc < (int)ks.spl_num_kpoints().local_size(); ikloc++)
+    for (int ikloc = 0; ikloc < ks.spl_num_kpoints().local_size(); ikloc++)
     {
-        int ik = (int)ks.spl_num_kpoints(ikloc);
+        int ik = ks.spl_num_kpoints(ikloc);
         auto kp = ks[ik];
         auto occupied_bands = kp->get_occupied_bands_list(kp->blacs_grid_slice().comm_col());
         add_k_point_contribution<ultrasoft_pseudopotential>(ks[ik], occupied_bands, pp_complex_density_matrix);
